Redish/requirements.txt
Django>=4.2
djangorestframework
django-environ
redis
django-redis